Markiplier Fan Bot
This program has the user input a YouTube Video URL. If it detects that the video was uploaded by "Markiplier", then it will borderline lock the user into watching the video until it ends. Otherwise, the program just quits. I'm slightly disappointed by the program because it doesn't completely lock your mouse, most ways to do that would mess up the while loop determining when the video ends. I could have uninstalled the mouse drivers, but that's kinda fucked up.